+++ Projects/Features/FeatureStudent/Sources/QR/Component/QRScannerRepresentable.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct QRScannerRepresentable: UIViewControllerRepresentable {
+    let onCodeScanned: (String) -> Void
+
+    func makeUIViewController(context: Context) -> QRScannerViewController {
+        let controller = QRScannerViewController()
+        controller.onCodeScanned = onCodeScanned
+        return controller
+    }
+
+    func updateUIViewController(_ uiViewController: QRScannerViewController, context: Context) {}
+}
