+++ Projects/Features/FeatureStudent/Sources/QR/Component/QRScannerViewController.swift
@@ -1,0 +1,116 @@
+import UIKit
+import AVFoundation
+
+final class QRScannerViewController: UIViewController, AVCaptureMetadataOutputObjectsDelegate {
+    var onCodeScanned: ((String) -> Void)?
+
+    private let captureSession = AVCaptureSession()
+    private let sessionQueue = DispatchQueue(label: "com.finda.qr.capture-session")
+    private var previewLayer: AVCaptureVideoPreviewLayer?
+    private var lastScannedCode: String?
+    private var isSessionConfigured = false
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .black
+        configureCameraSession()
+    }
+
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        previewLayer?.frame = view.bounds
+    }
+
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        if isSessionConfigured {
+            startSessionIfNeeded()
+        }
+    }
+
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        sessionQueue.async { [weak self] in
+            guard let self, self.captureSession.isRunning else { return }
+            self.captureSession.stopRunning()
+        }
+    }
+
+    private func configureCameraSession() {
+        switch AVCaptureDevice.authorizationStatus(for: .video) {
+        case .authorized:
+            setupSessionAndStart()
+        case .notDetermined:
+            AVCaptureDevice.requestAccess(for: .video) { [weak self] granted in
+                guard granted else { return }
+                DispatchQueue.main.async {
+                    self?.setupSessionAndStart()
+                }
+            }
+        default:
+            return
+        }
+    }
+
+    private func setupSessionAndStart() {
+        if isSessionConfigured {
+            startSessionIfNeeded()
+            return
+        }
+
+        guard let backCamera = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: .back) else {
+            return
+        }
+
+        do {
+            let input = try AVCaptureDeviceInput(device: backCamera)
+            if captureSession.canAddInput(input) {
+                captureSession.addInput(input)
+            }
+
+            let output = AVCaptureMetadataOutput()
+            if captureSession.canAddOutput(output) {
+                captureSession.addOutput(output)
+                output.setMetadataObjectsDelegate(self, queue: .main)
+                let availableTypes = output.availableMetadataObjectTypes
+                guard availableTypes.contains(.qr) else { return }
+                output.metadataObjectTypes = [.qr]
+            }
+
+            let preview = AVCaptureVideoPreviewLayer(session: captureSession)
+            preview.videoGravity = .resizeAspectFill
+            preview.frame = view.bounds
+            view.layer.insertSublayer(preview, at: 0)
+            previewLayer = preview
+
+            isSessionConfigured = true
+            startSessionIfNeeded()
+        } catch {
+            return
+        }
+    }
+
+    private func startSessionIfNeeded() {
+        sessionQueue.async { [weak self] in
+            guard let self, !self.captureSession.isRunning else { return }
+            self.captureSession.startRunning()
+        }
+    }
+
+    func metadataOutput(
+        _ output: AVCaptureMetadataOutput,
+        didOutput metadataObjects: [AVMetadataObject],
+        from connection: AVCaptureConnection
+    ) {
+        guard
+            let readableObject = metadataObjects.first as? AVMetadataMachineReadableCodeObject,
+            let code = readableObject.stringValue
+        else {
+            return
+        }
+
+        guard lastScannedCode != code else { return }
+        lastScannedCode = code
+        onCodeScanned?(code)
+    }
+}
