+++ Projects/Features/FeatureStudent/Sources/QR/QRScanView.swift
@@ -1,15 +1,18 @@
 import SwiftUI
-import AVFoundation
 
 public struct QRScanView: View {
+    // 최근에 읽은 QR 문자열(알럿 메시지에 표시)
     @State private var scannedURI: String = ""
+    // 스캔 완료 알럿 표시 상태
     @State private var isShowingScanAlert = false
 
     public init() {}
 
     public var body: some View {
         ZStack {
+            // UIKit 기반 QR 스캐너를 SwiftUI 뷰에 임베드
             QRScannerRepresentable { code in
+                // 알럿이 떠 있는 동안에는 중복 처리 방지
                 guard !isShowingScanAlert else { return }
                 scannedURI = code
                 isShowingScanAlert = true
@@ -21,131 +24,5 @@ public struct QRScanView: View {
         } message: {
             Text(scannedURI)
         }
-    }
-}
-
-private struct QRScannerRepresentable: UIViewControllerRepresentable {
-    let onCodeScanned: (String) -> Void
-
-    func makeUIViewController(context: Context) -> QRScannerViewController {
-        let controller = QRScannerViewController()
-        controller.onCodeScanned = onCodeScanned
-        return controller
-    }
-
-    func updateUIViewController(_ uiViewController: QRScannerViewController, context: Context) {}
-}
-
-private final class QRScannerViewController: UIViewController, AVCaptureMetadataOutputObjectsDelegate {
-    var onCodeScanned: ((String) -> Void)?
-
-    private let captureSession = AVCaptureSession()
-    private let sessionQueue = DispatchQueue(label: "com.finda.qr.capture-session")
-    private var previewLayer: AVCaptureVideoPreviewLayer?
-    private var lastScannedCode: String?
-    private var isSessionConfigured = false
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .black
-        configureCameraSession()
-    }
-
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-        previewLayer?.frame = view.bounds
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        if isSessionConfigured {
-            startSessionIfNeeded()
-        }
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        sessionQueue.async { [weak self] in
-            guard let self, self.captureSession.isRunning else { return }
-            self.captureSession.stopRunning()
-        }
-    }
-
-    private func configureCameraSession() {
-        switch AVCaptureDevice.authorizationStatus(for: .video) {
-        case .authorized:
-            setupSessionAndStart()
-        case .notDetermined:
-            AVCaptureDevice.requestAccess(for: .video) { [weak self] granted in
-                guard granted else { return }
-                DispatchQueue.main.async {
-                    self?.setupSessionAndStart()
-                }
-            }
-        default:
-            return
-        }
-    }
-
-    private func setupSessionAndStart() {
-        if isSessionConfigured {
-            startSessionIfNeeded()
-            return
-        }
-
-        guard let backCamera = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: .back) else {
-            return
-        }
-
-        do {
-            let input = try AVCaptureDeviceInput(device: backCamera)
-            if captureSession.canAddInput(input) {
-                captureSession.addInput(input)
-            }
-
-            let output = AVCaptureMetadataOutput()
-            if captureSession.canAddOutput(output) {
-                captureSession.addOutput(output)
-                output.setMetadataObjectsDelegate(self, queue: .main)
-                let availableTypes = output.availableMetadataObjectTypes
-                guard availableTypes.contains(.qr) else { return }
-                output.metadataObjectTypes = [.qr]
-            }
-
-            let preview = AVCaptureVideoPreviewLayer(session: captureSession)
-            preview.videoGravity = .resizeAspectFill
-            preview.frame = view.bounds
-            view.layer.insertSublayer(preview, at: 0)
-            previewLayer = preview
-
-            isSessionConfigured = true
-            startSessionIfNeeded()
-        } catch {
-            return
-        }
-    }
-
-    private func startSessionIfNeeded() {
-        sessionQueue.async { [weak self] in
-            guard let self, !self.captureSession.isRunning else { return }
-            self.captureSession.startRunning()
-        }
-    }
-
-    func metadataOutput(
-        _ output: AVCaptureMetadataOutput,
-        didOutput metadataObjects: [AVMetadataObject],
-        from connection: AVCaptureConnection
-    ) {
-        guard
-            let readableObject = metadataObjects.first as? AVMetadataMachineReadableCodeObject,
-            let code = readableObject.stringValue
-        else {
-            return
-        }
-
-        guard lastScannedCode != code else { return }
-        lastScannedCode = code
-        onCodeScanned?(code)
     }
 }
